Refactored state machine logic slightly
Removed StandaloneStateMachine as it directly contradicted the intended way of using the sate machines.

Removed class version of ManagedStateMachine as it was actually neater to use just the interface.

Made more functions virtual and added a Initialize function to StateManagers to (hopefully) explicitly tell the implementer to Initialize the states.

Moved AudioEventSettings and AudioEventSettingsReference into one file.